Fixed Exception
Fixed the exception that we encountered in class. Exception: "Cannot dynamically create an instance of type 'AppliedActivity3.Services.IDataStore'. Reason: Cannot create an instance of an interface.". The problem was that I accidently registered the interface with the dependency service in the App.xaml.cs instead of the Sqlite implementation.